working based off guided practice up to step 5

+++ server/index.js
@@ -23,7 +23,7 @@ app.use('/assets', express.static(path.join(__dirname, '../client/dist/assets'))
 // check if user is logged in
 const isLoggedIn = async(req, res, next)=> {
   try {
-    req.user = await findUserByToken(req.headers.authorization);
+    req.user = await findUserWithToken(req.headers.authorization);
     next();
   }
   catch(ex){
@@ -44,7 +44,8 @@ app.post('/api/auth/login', async(req, res, next)=> {
 // retrieves user information
 app.get('/api/auth/me', isLoggedIn, async(req, res, next)=> {
   try {
-    res.send(await findUserWithToken(req.headers.authorization));
+    res.send(req.user);
+    //res.send(await findUserWithToken(req.headers.authorization));
   }
   catch(ex){
     next(ex);
@@ -64,6 +65,11 @@ app.get('/api/users', async(req, res, next)=> {
 // retrieves favorites for a specifc user
 app.get('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
   try {
+    if(req.params.id !== req.user.id){
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
     res.send(await fetchFavorites(req.params.id));
   }
   catch(ex){
@@ -74,6 +80,11 @@ app.get('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
 // creates a favorite for a user
 app.post('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
   try {
+    if(req.params.id !== req.user.id){
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
     res.status(201).send(await createFavorite({ user_id: req.params.id, product_id: req.body.product_id}));
   }
   catch(ex){
@@ -84,6 +95,11 @@ app.post('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
 // deletes a favorite of a user
 app.delete('/api/users/:user_id/favorites/:id', isLoggedIn, async(req, res, next)=> {
   try {
+    if(req.params.userId !== req.user.id){
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
     await destroyFavorite({user_id: req.params.user_id, id: req.params.id });
     res.sendStatus(204);
   }
